types [nfc]: Demo using the new Outbox pm/stream type distinction
These two asserts just confirmed that the `type` and
`display_recipient` properties varied together in the expected way,
in order to assure Flow that after checking `type` we'd get the
right type of value out of `display_recipient`.

Now that we've expressed that tandem variation directly in the actual
types, we can drop the asserts.