example data: Use $Rest, not $Shape
$Shape is not very sound -- Flow will sometimes treat the result as if
it definitely has all the original object type's properties.

Instead, use `$Rest<Foo, { ... }>` here, just as we do for
`streamMessage` and others in this file.

In fact, as a bonus simplification, fuse this use of `$Rest` with an
inexact object type (for making everything optional) with the existing
use of `$Diff` with an exact object type (for excluding a specific
property), resulting in a single `$Rest` with a combined type.