OptionRow: Make height the same regardless of icon.
This makes the OptionRow always the same height, regardless of whether
there's an icon or not. As far as I can tell, this doesn't change any of
the current uses of OptionRow (since almost all of them currently use
icons).

The value of 56 was discovered by inspecting OptionRows with icons -
since RNVI uses font-size instead of width/height, there isn't a nice
way to get the height of an icon from its size.

This is in preparation for removing some icons from the settings screen.