deps: Sync react-dom's range with react's.
We don't actually use react-dom; it's there for a kind of silly
reason (see 62621ef8e and related history:
`git log --stat -p -G react-dom -- package.json`). But, as long as
we include it, we should have a sensible plan for keeping it
up-to-date. Might as well do that by having it track with react.