unread [nfc]: Add a getter, getUnreadCountForTopic.
This will let us get this information in a convenient way from code
that's entirely outside of the unreads code -- like this call site,
in `getTopicsForStream` which is a UI-oriented selector that exists
to drive the topics-list screen -- without having to dig into the
details of the data structures we use for the unreads data.

It's possible to get a similar effect already by invoking
`getUnreadCountForNarrow`.  But (a) that interferes with the caching
on that selector, which is really meant to be used only for the one
current narrow that's foregrounded in the UI; and (b) that requires
an entire `GlobalState` object, whereas we'd like to be able to get
this effect from code that has only some relevant pieces of the state,
in the spirit of `BackgroundData`.

Because we generally say "selector" to mean something that requires
the global state as its input, we file this function under the more
general term "getter".