messages: Don't hide muted messages in mentioned narrow.
Currently, we don't show muted messages in the mentioned narrow. This is
confusing and inconsistent with the webapp. This commit fixes that.

We should do a more thorough audit of this code, and compare it to the
webapp version. I haven't attempted to do that right now, since this
small tactical fix seems important and high-priority.

Related: #3472