notifications: Add more detail to "realm_uri ambiguous" error.
There are a lot of users seeing this - more than I would expect to have
multiple accounts on the same server. I want to collect this data so
that we can have a more complete picture of what's going on in these
cases.

These counts will tell us enough to determine whether people are ending
up somehow with duplicate accounts in our state, or on the other hand
have several accounts on one realm that actually seem to be distinct.