generateInboundEvents [nfc]: Add another TODO, following a relevant change.
The previous commit, where we removed an early return, worsened a
latent but tiny bug: some code that will sometimes send an
unnecessary 'read' event is now being run on changes to
`htmlPieceDescriptorsForShownMessages`. Before, it was being skipped
by the early return [1].

The 'read' event, we expect, won't have a really noticeable effect
beyond a potential slight slowdown; it tells the WebView that some
messages -- which the WebView already knows are read -- are read.

[1] https://chat.zulip.org/#narrow/stream/243-mobile-team/topic/message-list.20diffing/near/1222510