AccountPickScreen: Simplify reasoning for `canGoBack`.
`canGoBack` is false just when the navigation stack is empty, and
true otherwise.

This should match the previous behavior: the stack will be non-empty
only when we've gotten here from the "switch accounts" button, so
we'll pass `true` for `canGoBack`, just as before.

(For the logout case, we'll always do a "reset" action instead of a
"push" action to get here, meaning the stack will be empty.)

canGoBack isn't yet documented, but we hope it will be soon:
  https://github.com/react-navigation/react-navigation.github.io/issues/724