reducers types [nfc]: Add an annotation.
Without this, when I try upgrading to Flow v0.137, for the upcoming
RN v0.64 upgrade, I get this error:

```
Error --------------------------------------------------------------------------------------- src/boot/reducers.js:73:33

Cannot call `maybeLogSlowReducer` with `key` bound to `key` because type variable `Key` [1] cannot escape from the scope
in which it was defined [2] (try adding a type annotation to `key` [3]). [escaped-generic]

   src/boot/reducers.js:73:33
   73|     maybeLogSlowReducer(action, key, startMs, endMs);
                                       ^^^

References:
   src/boot/reducers.js:43:8
   43|   key: Key,
              ^^^ [1]
   src/boot/reducers.js:42:23
   42| function applyReducer<Key: $Keys<GlobalState>, State>(
                             ^^^ [2]
   src/boot/reducers.js:36:38
   36| function maybeLogSlowReducer(action, key, startMs, endMs) {
                                            ^^^ [3]
```

I haven't reasoned all of the way through it yet, but the annotation
seems harmless and correct.