sharing: Set appropriate filename of shared content.
Previously filename was the last element of
sharedFileUrl.split('/') or sharedImageUrl.split('/').

This did not always correspond to the actual filename, sometimes
this element did not contain a file extension making the uploaded
data to be stored as a binary file. This made shared images fail
to render and on downloading the shared content it was downloaded
as a binary file.

This commit uses ContentResolver to find the actual name of file
being shared and uses that to identify the shared file.

`getFileName` introduced in this commit is based on:
- https://developer.android.com/training/secure-file-sharing/retrieve-info.html

the query method used in `getFileName` has the potential to return
null in case of invalid url, we are handling it by setting our own
name "unknown.<extension>" <extension> being the one that was found
by contentResolver.getType, if that fails too then we just set it
to `.bin`.