api errors: Throw an exception to reach logging code.
This is NFC in the case where we reach `makeErrorFromApi`
and throw an exception it returns.

If there's an exception somewhere else -- which would be a bug in
this code -- then this change means that we now reach the logging
code that got moved inside this new `catch` block.  That seems
all to the good.

And in fact, one of our tests demonstrates that case!  It does so
by artificially making `fetch` itself fail.  That does seem like a
bug (an existing one) in this code; we'll be fixing it later in
this series.