i18n: Sync translations from Transifex
Thanks as always to our kind volunteer translators.

Note: The recently-added strings are synced across languages in this
commit, instead of in a separate "Sync recently-added message strings"
commit. That's because I ran `tools/tx-sync` twice after that string
was added, and I kept the result from the second time. This was a
small mistake that I don't expect to cause real problems, but it does
explain what happened with that string.