mute [nfc]: Make shouldBeMuted's arguments required.
It doesn't make any sense to call this function without passing
these arguments.

We were leaving one of them off in some test cases.  Perhaps in some
sense it's more convenient that way in the tests.  But really the
main effect is to make those test cases confusing about what the
interface to this function really is -- how does it learn what topics
are muted?  Better to spend the four characters to be explicit.