msglist [nfc]: Simplify getMessageUpdateStrategy control flow (4/6).
Here there are no duplicate rechecks of these conditions to delete,
so the changes are very simple and we do several of them in one go.
We're just writing the definitions of these conditions in the same
order as we actually consult them in, and returning immediately
once we determine the result.