deps: Upgrade jetifier to 2.0.0, the latest.
Here's the description from the changelog [1]:

> Exact same as 1.6.7 was, just as a breaking change semver number
>
> docs(CHANGELOG): prepare release 2.0.0 f483026
> BREAKING: note that jetifier is deprecated, warn on translations
>   699094b

It looks like in 1.6.7 the tool started giving an important
deprecation warning about itself. But then the warning was taken out
in 1.6.8 because it was deemed to be a breaking change that deserved
a new major version. 2.0.0 was that new major version.

Right now, for us, the warning looks like this:

```
Jetifier is deprecated. Future versions of react-native CLI may not
  run it by default.
Jetifier found 1372 file(s) to forward-jetify. Using 12 workers...
Jetifier: propose an AndroidX conversion PR to this repository:
  node_modules/react-native-photo-view/android/src/main/java/com/reactnative/photoview/ImageEvent.java
Jetifier: propose an AndroidX conversion PR to this repository:
  node_modules/react-native-photo-view/android/src/main/java/com/reactnative/photoview/PhotoView.java
```

So there are just two files in our dependency tree that still need to
move to AndroidX, both in react-native-photo-view.  We have an issue
already (#4217) for using something more maintained than that library,
which should take care of those.

Anyway, take this upgrade, in the spirit of not ignoring warnings
that we should act on. And it's just four lines of output.

Also I guess because we'll want `yarn upgrade` to automatically give
us any new, non-breaking changes in the tool.

[1] https://github.com/mikehardy/jetifier/releases