i18n: Use Transifex's names for the files synced from Transifex.
This reverts part of the changes from commit 0d0d898b7,
"i18n: Use valid language tags (with `-` instead of `_`)."

Specifically:

 * We keep using BCP 47 language tags, our preferred form, for
   referring to the languages inside our own code, and in particular
   for recording the user's preference in `state.settings`.

 * But for the files `static/translations/messages_*.json`, which are
   frequently regenerated from the data in Transifex by the Transifex
   client `tx` (as invoked by `tools/tx-sync`), we revert to the
   language identifiers that Transifex uses.

As things stand, `tools/tx-sync` downloads the data for a language
like pt-PT to a file named like messages_pt_PT.json, which is a file
that doesn't exist in the tree.  In order to properly complete the
sync, the file has to manually be renamed.

The alternative way to fix this would be to add some code to
`tools/tx-sync` for it to rename those files over after each time it
runs `tx pull`.  That would work but seems like avoidable complexity.
The Transifex names are unavoidably something maintainers will
sometimes interact with anyway, because they appear in the Transifex
web interface.