deps: Run `yarn upgrade`.
On `pod install`, I got this error:

```
[!] CocoaPods could not find compatible versions for pod "Sentry":
  In snapshot (Podfile.lock):
    Sentry (= 7.0.0)

  In Podfile:
    RNSentry (from `../node_modules/@sentry/react-native`) was resolved to 2.6.2, which depends on
      Sentry (= 7.1.4)

Specs satisfying the `Sentry (= 7.0.0), Sentry (= 7.1.4)` dependency were found, but they required a higher minimum deployment target.
```

It went away after I ran `pod update Sentry`, as in 4dd471c31, and
the error went away.

I'm not sure if it was trying to say that the Sentry pod has a
higher minimum deployment target *than the one we declare*. In any
case, that'd be the important thing to find out. Checking, it looks
like it doesn't: `platforms.ios` in its Podspec is "9.0" [1]. Our
deployment target is 12.1; that appears at the top of our Podfile.
So there should be no incompatibility with any supported devices.

This also includes a recent update to ImmutableJS, following a
three-year period of no updates. :tada:

[1] https://github.com/CocoaPods/Specs/blob/master/Specs/a/b/d/Sentry/7.1.4/Sentry.podspec.json#L13