ZulipBanner [nfc]: Better support single-row layout (2/3).
To provide the vertical distance of 12px between the text row and
the buttons row, declared in the two-line-text, two-buttons
illustration in the spec [1], the status quo chooses the margin
placement arbitrarily between `textRow` and `buttonsRow`. So, moving
the margin from one to the other is fine and NFC.

But it'll help us better approximate centering the text and buttons
vertically, when we start allowing the text and buttons to share a
single row when there's room.

[1] https://material.io/components/banners#specs