redux: Store userId in state.accounts
Earlier we used to identify users by their emails, making it difficult
to keep track of which user to switch to if multiple accounts are logged
in to the same org when opening the app via a notification.

Since server version 2.1-dev-540-g447a517e6f the server sends the
user ID of the relevant account in the notification. Making it sensible
to store the userId in the redux state.

Fixes: #4951