ios/Podfile: Add instructions from upstream for enabling Hermes on iOS.
Done to follow the template-app change in
facebook/react-native@eb85d1dbd.

Done after the RN v0.64 upgrade because Hermes on iOS is advertised
as new in v0.64.

We don't actually intend to use Hermes at all until RN v0.65, when
it gains support for Intl when used on Android. Intl support on iOS
is still under discussion, and they don't have a timeline yet:
  https://github.com/facebook/hermes/issues/23#issuecomment-912017440