outbox: Add a stream_id property
Compare a1fad7ca9, which did the same for the sender_id property.

We actually did part of this a few months ago, in fce3e8b3d in #4667.
That commit added it to the type, but we hadn't yet taken the step of
supplying this property in any actual Outbox values.

Also add back to a TODO comment a bit that got dropped in 8171f2d7b.

This makes #3998 mostly complete: the remaining steps will be to get a
release with this change out, wait a few weeks, then make `stream_id`
required with a migration to drop `Outbox` values that lack it, just as
we did for `sender_id` before.

Issue: #3998