compose [nfc]: Rename and document showComposeBoxOnNarrow
The name `canSendToNarrow` was misleading: we never actually send
to a whole-stream narrow, but this function returns true for them.
The real point is that we show the compose box when you're looking
at that narrow.

(When you do look at a whole-stream narrow and we show the compose
box, it then offers a text input for a topic, and ensures through
its `getDestinationNarrow` that we send to the chosen topic's
specific narrow rather than the whole stream.)