webview html [nfc]: Clarify when we show any recipient headers
This local was `showHeader`, but it's really not about any one
header; rather it's about whether this is a narrow where we're
going to show these headers in general at all.

Moreover the specific kind of header we're talking about here is a
recipient header.  No need to risk confusion with, say, where we
show the sender at the top left of some messages.