webview html [nfc]: Try to clarify show-sender logic
The main change here is to flip the sense of this boolean local.
This way around the logic feels easier to understand, to me: we
show the sender if this, or that, or the other.  The old code feels
like it requires unpacking some double negatives to understand.

The actual expression isn't quite as nice as the explanation in
the comment.  We'll fix that up in the next few commits.