account: Deprecate combined tryGetAuth, adding tryGetThisAuth
Much like tryGetActiveAccount and getActiveAccount, this function's
callers are a mix of global and per-account.

The first step of what this does is to go from global to per-account
state; we just introduced tryGetActiveAccountState for that.  Make
a function for specifically the second step, getting an `Auth` from
a per-account state, and deprecate this one.

After migrating this function's callers to all be explicit about
whether they're global or per-account, we'll rename the new
per-account version to the old name.