reducers [nfc]: Apply per-account reducers only on per-account actions
This is NFC because these sub-reducers already ignore any actions
that aren't in PerAccountAction.  (Flow won't check this for us;
but I just grepped for for `case` labels in these reducers' files,
piped that to `sort -u`, and confirmed that all the results are
as expected.)

Here "ignore" means they return the state they're given, but after
applying JS optional-argument semantics to it -- so when passed
`undefined`, they return their respective `initialState` values.