msglist [nfc]: Use assumeSecretly, needed for `connect` to give PerAccountState
This is a per-account component, and this `connect` call will soon
start providing only a per-account state.  We'll need to fix one way
or another the way it gets these bits of global state (all of which
are perfectly legitimate for per-account code to be using), but for
now just mark that.