config [nfc]: Inline away enableWebViewErrorDisplay
Our use of this config variable has been to control whether we show
the error display in the WebView. As Greg points out [1], there
isn't really any point in having that happen in config.js, which is
pretty far away from the code that actually controls whether the
error display is shown.

While we're at it, rename the variable in js.js to just
`isDevelopment`. That's a very natural bit of background info to
want to know (enableWebViewErrorDisplay is unnecessarily specific),
and we may have another use case for it soon.

[1] https://github.com/zulip/zulip-mobile/pull/5027#discussion_r718967861