react-redux types: withHaveServerDataGate needn't care about dispatch type
This component discards the `dispatch` prop it gets, so it doesn't
really care about its type and can say it merely expects a `mixed`.
This will let us change the `dispatch` type that gets passed here
without having to make a meaningless edit at this line.

Previously we couldn't write `mixed` here because of a bug in our
`OwnProps` type, which is used in our `connect`.  We've just fixed
that by using our `BoundedDiff` in place of `$Diff`.