ios: Run `pod update DoubleConversion` and `pod update glog`
As, apparently, we should have done in the recent RN v0.64 upgrade,
4a5780d65. I verified that `pod update DoubleConversion` and
`pod update glog` made these changes at that upgrade commit, but
these commands made no changes just before the upgrade commit.

For other people stumbling on this, see
  https://github.com/facebook/react-native/issues/31121#issuecomment-873901644.

This looks like another case of what Greg described on CZO [1]:

> Yeah, OK, so one thing that may be going on here is:
> * [CocoaPods] sees a dependency for a certain pod at a certain
>   version
> * It already has it at that version, so it says great, I'm done,
>   that's it
> * ... But actually there are different versions floating around
>   that have the same version number, and the dependency meant to
>   refer to the other one.

For DoubleConversion, I verified experimentally that
facebook/react-native@216037757 had the offending (one-line) change
to the podspec without the version being bumped. I was able to
revert that specific change in our node_modules/react-native and see
the Podfile.lock change go away when I ran
`pod update DoubleConversion` again.

No such luck with glog; I don't know why. That same commit made a
similar-looking change to glog's podspec, but I got a different hash
when I reverted that and ran `pod update glog`. That's the only
commit that touched glog's podspec that was new in RN v0.64, though,
so I'm not really sure what's going on. Maybe the podspec pulls in
some code, and that code changed.

Anyway, `pod update glog` is something we might reasonably have done
as part of the RN upgrade commit, and we'd have gotten the current
result if we had done so.

Note that there's a *different* known issue with unstable hashes in
the Podfile.lock. The symptoms are that `FBReactNativeSpec` will
usually get different hashes on different computers because the
output of some `pod` code includes absolute paths, and that ends up
affects the hashes. See facebook/react-native@bdfe2a517 for where
this is supposedly fixed in RN v0.65, and
https://github.com/zulip/zulip-mobile/pull/4986#discussion_r714198594
for a real-world example of us stumbling on it.

[1] https://chat.zulip.org/#narrow/stream/243-mobile-team/topic/ios.20build.3A.20double-conversion/near/913869