app bar: Use flex properly in titles, preventing collisions
Before this fix, if you visit a stream or topic narrow with a very
long stream name, the text of the name will run over on top of the
button icons that go after it.

Similarly, if you visited a narrow that uses TitleSpecial (so
all-messages, all-PMs, starred, or @-mentions) and the title we use
to describe that narrow was super long (due to translation, perhaps),
the title would run off the end without a proper ellipsis.

Fix those by using Flexbox properly so that it can see this is the
component to put the residual width into, after the fixed amounts for
buttons and padding and so on.

All other "Title*" components already get this right.