app bar: Make the padding in ModalNavBar halfway reasonable
We'll fix this up to more precisely match the intended Material
Design spec in a bit.

But first, just fix up two aspects that were totally absurd:
 * NAVBAR_SIZE, which is 58px, is vastly too much right margin here.
   Make it 16px, the same as the other margin here.
 * On the other hand, when there's no back button, we were using
   no right margin at all.  Instead, make it symmetric.

We leave out a left margin in the canGoBack case because the back
button provides its own padding.  The amount there isn't quite right
(that's part of what we'll fix up in a bit) but doesn't look
obviously wrong in itself.