msglist: Fix bug breaking heavily formatted links (like bold+italic)
The TODO here was onto something, it turns out: Before this change,
if you tapped a link that's both bold and italic, in a message,
nothing would happen. Now, the link opens, as you'd expect.

While we don't know exactly how many levels of nesting the server
might give us, it's convenient and it seems sensible enough to
handle an arbitrarily large number.

Also, move this case downward, as Greg says, "so that if we have an
`a` with a lot of stuff inside it, and we have something more
specific in there, the more specific thing works"; see more detail
at
  https://github.com/zulip/zulip-mobile/pull/5048#discussion_r728522394.